rename : move files (#22)
